Migrate temp file writing to helper module
Will be used by a client test in the future.
